Переработаны арифметические опреции, добавлена полноценная поддержка многомерных массивов. Необходимо перерабоать так же все остальные операции и рассмотреть вариант замены статических массивов MyVariant::elements на динамические

+++ definitions.h
@@ -13,7 +13,7 @@
 #include <QStringList>
 #include <QVector>
 
-typedef long double real_type;
+typedef double real_type;
 
 void reportError(const QString &ss);
 void reportRuntimeError(const QString& ss);
